+++ .travis.yml
@@ -5,6 +5,6 @@ cache: yarn
 install:
   - yarn
 script:
-  - npm test
+  - yarn test
 after_script:
-  - npm run coveralls
+  - yarn run coveralls
