+++ README.md
@@ -4,7 +4,7 @@
 [![Coverage Status](https://coveralls.io/repos/github/jorritdenbroeder/atmention/badge.svg?branch=master)](https://coveralls.io/github/jorritdenbroeder/atmention?branch=master)
 [![npm version](https://badge.fury.io/js/atmention.svg)](https://badge.fury.io/js/atmention)
 
-Textarea with @mention autocomplete
+Textarea with @mention autocomplete for AngularJS
 
 ## Installation (AngularJS)
