+++ packages/atmention-angularjs-demo/app/app.module.js
