+++ packages/atmention-angularjs-demo/app/root.component.js
