+++ packages/atmention-angularjs-demo/index.js
