+++ packages/atmention-angularjs-demo/styles/index.scss
