+++ packages/atmention-angularjs/index.js
