+++ packages/atmention-angularjs/karma.conf.js
