+++ packages/atmention-angularjs/package.json
@@ -14,6 +14,6 @@
   "author": "Jorrit den Broeder <jorrit@parksound.nl>",
   "license": "ISC",
   "dependencies": {
-    "atmention": "^0.3.1"
+    "atmention-core": "^0.3.1"
   }
 }
