+++ packages/atmention-angularjs/src/atmention.filter.js
