+++ packages/atmention-angularjs/src/atmention.module.js
