+++ packages/atmention-angularjs/src/input.component.js
