+++ packages/atmention-angularjs/src/item-template-directives.js
