+++ packages/atmention-angularjs/src/render.component.js
