+++ packages/atmention-angularjs/src/styles/index.scss
