+++ packages/atmention-angularjs/test/index.js
