+++ packages/atmention-core/index.js
