+++ packages/atmention-core/karma.conf.js
