+++ packages/atmention-core/src/constants.js
