+++ packages/atmention-core/src/format.js
