+++ packages/atmention-core/src/util.js
